consumer/requirements.txt
kafka-python
numpy
python-dotenv